Refactor frame and log handling in RustSynth
- Removed unused CStr imports in frame module.
- Enhanced safety documentation for FrameContext and LogHandle creation from raw pointers.
- Updated string retrieval methods in Map to simplify signature.
- Improved error handling in Plugin invocation, adding specific error messages.
- Adjusted output retrieval methods in Environment to accept i32 directly instead of casting.
- General code cleanup and consistency improvements across modules.

+++ rspipe/src/main.rs
@@ -267,7 +267,7 @@ fn process_frames_concurrent(
                 Ok(frame) => Ok(frame),
                 Err(e) => Err(format!("Frame error: {}", e)),
             };
-            tx_clone.send((n as usize, result_owned)).unwrap();
+            tx_clone.send((n, result_owned)).unwrap();
             *pending_clone.lock().unwrap() -= 1;
         });
     }
@@ -297,7 +297,7 @@ fn process_frames_concurrent(
                                 Ok(frame) => Ok(frame),
                                 Err(e) => Err(format!("Frame error: {}", e)),
                             };
-                            tx_clone.send((n as usize, result_owned)).unwrap();
+                            tx_clone.send((n, result_owned)).unwrap();
                             *pending_clone.lock().unwrap() -= 1;
                         });
                     }
